Microeconomics Basic Game Design
Graded project due after completion of an online game design course on Coursera.
Monopoly Pizza Shop, demonstrates the concept of monopoly pricing. The player controls the price of pizzas. As price increases, the number of customers falls, but revenue per customer rises. As price decreases, more customers appear, but profit may shrink due to higher costs. By experimenting with different prices, the player finds the output where profit is maximized. 
To play the game, you are the owner of the monopoly pizza shop in town. Simply use the up and down arrow keys to toggle and control the price of your pizza.
You win the game when you find the profit maximizing price.